+++ tricorder/src/Tricorder/Effects/GhciSession.hs
@@ -15,11 +15,11 @@ module Tricorder.Effects.GhciSession
 
 import Atelier.Effects.Conc (Conc)
 import Atelier.Effects.File (File)
+import Atelier.Effects.Process (Process)
 import Atelier.Effects.Timeout (Timeout)
 import Data.Default (def)
 import Effectful
     ( Effect
-    , IOE
     , Limit (..)
     , Persistence (..)
     , UnliftStrategy (..)
@@ -105,7 +105,7 @@ runGhciSession
        , Conc :> es
        , Concurrent :> es
        , File :> es
-       , IOE :> es
+       , Process :> es
        , Timeout :> es
        )
     => Eff (GhciSession : es) a -> Eff es a